Split wizard into separate base class.
